toolkit/mozapps/downloads/tests/chrome/test_bug_412360.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * This tests bug 412360, which caused problems when trying to save javascript
 * URIs.  This is not an xpcshell unit test because it triggers an assertion,
 * which causes xpcshell test cases to fail.
-->

<window title="Download Manager Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

Components.utils.import("resource://gre/modules/XPCOMUtils.jsm");

let didFail = false;
var file;

let prefBranch = Cc["@mozilla.org/preferences-service;1"].
                 getService(Ci.nsIPrefBranch);

let factory = {
  createInstance: function(aOuter, aIid) {
    if (aOuter != null)
      throw Components.results.NS_ERROR_NO_AGGREGATION;
    return promptService.QueryInterface(aIid);
  }
};

let promptService = {
  QueryInterface: XPCOMUtils.generateQI([Ci.nsIPromptService]),
  alert: function() {
    ok(didFail, "javascript: uri failed and showed a message");
    file.remove(false);

    cleanUpFactory();
    SimpleTest.finish();
  }
};

Components.manager.QueryInterface(Ci.nsIComponentRegistrar).
registerFactory(Components.ID("{6cc9c9fe-bc0b-432b-a410-253ef8bcc699}"),
  "PromptService", "@mozilla.org/embedcomp/prompt-service;1",
  factory
);

function cleanUpFactory() {
    // Unregister the factory so we do not leak
    Components.manager.QueryInterface(Ci.nsIComponentRegistrar).
    unregisterFactory(
      Components.ID("{6cc9c9fe-bc0b-432b-a410-253ef8bcc699}"),
      factory
    );
}

function test()
{
  var dmui = getDMUI();
  if (!dmui) {
    cleanUpFactory();
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  let dm = Cc["@mozilla.org/download-manager;1"].
           getService(Ci.nsIDownloadManager);
  let db = dm.DBConnection;

  // Empty any old downloads
  db.executeSimpleSQL("DELETE FROM moz_downloads");

  let stmt = db.createStatement(
    "INSERT INTO moz_downloads (source, target, state, endTime) " +
    "VALUES (?1, ?2, ?3, ?4)");

  try {
    // Saving javascript URIs doesn't work
    stmt.bindStringParameter(0, "javascript:5");

    // Download to a temp local file
    file = Cc["@mozilla.org/file/directory_service;1"].
           getService(Ci.nsIProperties).get("TmpD", Ci.nsIFile);
    file.append("javascriptURI");
    file.createUnique(Ci.nsIFile.NORMAL_FILE_TYPE, 0666);
    stmt.bindStringParameter(1, Cc["@mozilla.org/network/io-service;1"].
      getService(Ci.nsIIOService).newFileURI(file).spec);

    // Start it as canceled
    stmt.bindInt32Parameter(2, dm.DOWNLOAD_CANCELED);

    stmt.bindInt32Parameter(3, Date.now() * 1000);

    // Add it!
    stmt.execute();
  }
  finally {
    stmt.finalize();
  }

  let listener = {
    onDownloadStateChange: function(aState, aDownload)
    {
      switch (aDownload.state) {
        case dm.DOWNLOAD_FAILED:
          // Remember that we failed for the prompt service
          didFail = true;

          ok(true, "javascript: uri failed instead of getting stuck");
          dm.removeListener(listener);
          break;
      }
    }
  };
  dm.addListener(listener);

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();


  let os = Cc["@mozilla.org/observer-service;1"].
           getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testObs = {
    observe: function(aSubject, aTopic, aData)
    {
      if (aTopic != DLMGR_UI_DONE)
        return;

      os.removeObserver(testObs, DLMGR_UI_DONE);

      SimpleTest.waitForFocus(function () {
        let win = aSubject;

        // Down arrow to select the download
        synthesizeKey("VK_DOWN", {}, win);

        // Enter key to retry the download
        synthesizeKey("VK_ENTER", {}, win);
      }, aSubject);
    }
  };

  // Register with the observer service
  os.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
